Invoice service initial implementation

+++ src/main/kotlin/com/adrrriannn/invoice/manager/dto/InvoiceItemDto.kt
@@ -1,5 +1,5 @@
 package com.adrrriannn.invoice.manager.dto
 
-class InvoiceItemDto(val id: Int, val description: String, val amount: Double) {
+class InvoiceItemDto(val id: Long, val description: String, val amount: Double) {
 
 }
