+++ src/main/kotlin/com/adrrriannn/invoice/manager/service/InvoiceService.kt
@@ -4,5 +4,5 @@ import com.adrrriannn.invoice.manager.dto.InvoiceDto
 
 interface InvoiceService {
     fun createInvoice(invoiceDto: InvoiceDto): InvoiceDto
-    fun get(id: Int): InvoiceDto
+    fun get(id: Long): InvoiceDto
 }
